correcting Meeting
Date.java and SerialDate.java, as a consequence are serialazable but
Firebase doesn't handle this class anymore. So now the dates are
converted into long to fit firebase. It indeed worked so far with date
but the logcat was continuously filled with error message. This cache is
only 256KB large so it was quickly full and made the app crashes.